fix(tours): coherence total_eqp fractionnaire (app chauffeur) + tour_type tolerant
Audit des effets de bord des modifs du jour :

- schemas/mobile.py : DriverTourRead.total_eqp et AvailableTourRead.total_eqp
  etaient int -> 500 potentiel cote app chauffeur depuis le passage de
  total_eqp en numeric (EQP fractionnaire). Passes en float.
- schemas/tour.py : TourBase.tour_type rendu optionnel (TourType | None) pour
  tolerer une eventuelle ligne non retro-remplie en lecture (evite un 500 si
  le backfill n'a pas tourne). Defaut LIVRAISON a la creation, le front traite
  null comme LIVRAISON.

+++ Chaos_Route/backend/app/schemas/mobile.py
@@ -198,7 +198,7 @@ class DriverTourRead(BaseModel):
     delivery_date: str | None = None
     departure_time: str | None = None
     return_time: str | None = None
-    total_eqp: int | None = None
+    total_eqp: float | None = None  # EQP fractionnaire (volumes injectés)
     status: str
     base_code: str | None = None
     base_name: str | None = None
@@ -217,7 +217,7 @@ class AvailableTourRead(BaseModel):
     code: str
     delivery_date: str | None = None
     departure_time: str | None = None
-    total_eqp: int | None = None
+    total_eqp: float | None = None  # EQP fractionnaire (volumes injectés)
     stops_count: int = 0
     driver_name: str | None = None
     vehicle_code: str | None = None
